refactor: Extract LibraryTab as separate component
