+++ .trellis/p/P-implement-agent-management/P-implement-agent-management.md
@@ -202,7 +202,8 @@ affectedFiles:
   apps/desktop/src/components/settings/agents/index.ts: Added export for
     RoleSelect component; Added ModelSelect and PersonalitySelect to barrel
     exports; Added DefaultsTab to barrel exports to maintain consistent import
-    patterns across the codebase.
+    patterns across the codebase.; Added export for LibraryTab component to
+    barrel export file for consistent module access.
   apps/desktop/src/components/settings/agents/__tests__/RoleSelect.test.tsx:
     Created comprehensive unit tests covering all states, functionality,
     accessibility, edge cases, and component behavior with 100% test coverage
@@ -243,7 +244,10 @@ affectedFiles:
     component - removed useAgentSearch hook, search imports (Search, X icons),
     search UI components (search bar, clear button, loading indicator, results
     count), search-specific empty states, and simplified logic to use direct
-    agents array instead of filteredAgents
+    agents array instead of filteredAgents; Removed LibraryTab component
+    definition, AgentGrid component definition, AgentGridProps interface, and
+    mock agents data. Added import for LibraryTab from './LibraryTab'. Cleaned
+    up imports that are no longer needed.
   apps/desktop/src/components/settings/agents/__tests__/AgentsSection.test.tsx:
     Updated unit tests to reflect the new two-tab structure by removing all
     template-related test assertions and expectations. Updated test descriptions
@@ -256,6 +260,11 @@ affectedFiles:
     AgentsSection.tsx. Includes all configuration controls, preview panel,
     accessibility features, and JSDoc documentation. Component maintains exact
     same functionality as original inline version.
+  apps/desktop/src/components/settings/agents/LibraryTab.tsx: Created new
+    component file containing LibraryTab component, AgentGrid component, and
+    AgentGridProps interface extracted from AgentsSection.tsx. Includes all
+    required imports, mock agent data, and full functionality for displaying and
+    managing agent library.
 log: []
 schema: v1.0
 childrenIds:
