+++ .trellis/p/P-implement-agent-management/e/E-ui-components/f/F-section-components-update/F-section-components-update.md
@@ -1,7 +1,7 @@
 ---
 id: F-section-components-update
 title: Section Components Update
-status: in-progress
+status: done
 priority: medium
 parent: E-ui-components
 prerequisites:
@@ -21,7 +21,10 @@ affectedFiles:
     component - removed useAgentSearch hook, search imports (Search, X icons),
     search UI components (search bar, clear button, loading indicator, results
     count), search-specific empty states, and simplified logic to use direct
-    agents array instead of filteredAgents
+    agents array instead of filteredAgents; Removed LibraryTab component
+    definition, AgentGrid component definition, AgentGridProps interface, and
+    mock agents data. Added import for LibraryTab from './LibraryTab'. Cleaned
+    up imports that are no longer needed.
   apps/desktop/src/components/settings/agents/__tests__/AgentsSection.test.tsx:
     Updated unit tests to reflect the new two-tab structure by removing all
     template-related test assertions and expectations. Updated test descriptions
@@ -35,13 +38,21 @@ affectedFiles:
     accessibility features, and JSDoc documentation. Component maintains exact
     same functionality as original inline version.
   apps/desktop/src/components/settings/agents/index.ts: Added DefaultsTab to
-    barrel exports to maintain consistent import patterns across the codebase.
-log: []
+    barrel exports to maintain consistent import patterns across the codebase.;
+    Added export for LibraryTab component to barrel export file for consistent
+    module access.
+  apps/desktop/src/components/settings/agents/LibraryTab.tsx: Created new
+    component file containing LibraryTab component, AgentGrid component, and
+    AgentGridProps interface extracted from AgentsSection.tsx. Includes all
+    required imports, mock agent data, and full functionality for displaying and
+    managing agent library.
+log:
+  - "Auto-completed: All child tasks are complete"
 schema: v1.0
 childrenIds:
+  - T-extract-defaultstab-as
   - T-extract-librarytab-as
   - T-remove-search-functionality
-  - T-extract-defaultstab-as
   - T-remove-templates-tab-and
 created: 2025-08-19T16:01:11.431Z
 updated: 2025-08-19T16:01:11.431Z
