+++ .trellis/p/P-implement-agent-management/e/E-ui-components/f/F-section-components-update/t/closed/T-extract-librarytab-as.md
@@ -1,13 +1,32 @@
 ---
 id: T-extract-librarytab-as
 title: Extract LibraryTab as separate component file
-status: open
+status: done
 priority: medium
 parent: F-section-components-update
 prerequisites:
   - T-remove-search-functionality
-affectedFiles: {}
-log: []
+affectedFiles:
+  apps/desktop/src/components/settings/agents/LibraryTab.tsx: Created new
+    component file containing LibraryTab component, AgentGrid component, and
+    AgentGridProps interface extracted from AgentsSection.tsx. Includes all
+    required imports, mock agent data, and full functionality for displaying and
+    managing agent library.
+  apps/desktop/src/components/settings/agents/AgentsSection.tsx:
+    Removed LibraryTab component definition, AgentGrid component definition,
+    AgentGridProps interface, and mock agents data. Added import for LibraryTab
+    from './LibraryTab'. Cleaned up imports that are no longer needed.
+  apps/desktop/src/components/settings/agents/index.ts: Added export for
+    LibraryTab component to barrel export file for consistent module access.
+log:
+  - Successfully extracted LibraryTab component from AgentsSection.tsx into its
+    own separate file. The extraction included moving the AgentGrid component
+    and AgentGridProps interface since they are only used by LibraryTab. All
+    functionality has been preserved including the simplified implementation
+    without search functionality, responsive grid layout, keyboard navigation,
+    empty state handling, and accessibility features. The extraction improves
+    code organization and maintainability while following the existing patterns
+    in the codebase.
 schema: v1.0
 childrenIds: []
 created: 2025-08-19T19:26:29.520Z
