+++ apps/desktop/src/components/settings/agents/index.ts
@@ -3,6 +3,7 @@ export { AgentsSection } from "./AgentsSection";
 export { DefaultsTab } from "./DefaultsTab";
 export { EmptyLibraryState } from "./EmptyLibraryState";
 export { EmptyTemplatesState } from "./EmptyTemplatesState";
+export { LibraryTab } from "./LibraryTab";
 export { ModelSelect } from "./ModelSelect";
 export { PersonalitySelect } from "./PersonalitySelect";
 export { RoleSelect } from "./RoleSelect";
